chore: Extract util function
